+++ tests/cpp/test_vec4_operations.cpp
@@ -1,6 +1,7 @@
 #include <catch2/catch.hpp>
 #include <cmath>
-#include <loco/math/all.hpp>
+
+#include <loco/math/vec4_t_impl.hpp>
 
 /*
  * @todo(wilbert): replace GENERATE of fixed values with random values + seed
@@ -133,8 +134,8 @@ TEMPLATE_TEST_CASE("Vector4 class (vec4_t) core Operations", "[vec4_t][ops]",
         auto scale = GENERATE(as<T>{}, 0.25, 0.50, 0.75, 1.0);  // NOLINT
 
         Vector4 v(val_x, val_y, val_z, val_w);
-        auto v_1 = scale * v;
-        auto v_2 = v * scale;
+        auto v_1 = static_cast<double>(scale) * v;
+        auto v_2 = v * static_cast<double>(scale);
 
         REQUIRE(std::abs(v_1.x() - (val_x * scale)) < EPSILON);
         REQUIRE(std::abs(v_1.y() - (val_y * scale)) < EPSILON);
